backend/python-endpoints/requirements.txt
Flask==1.0.2
opencv-python
numpy
scipy
pandas
google-cloud-vision
google-cloud-storage
functions-framework
h5py < 3.0.0
Pillow